pyproject.toml
[tool.poetry]
name = "home-security"
version = "0.1.0"
description = "An AI Application for CCTV intrusion detection"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
protobuf = "3.19.5"
pillow = "9.2.0"
numpy = "^1.26.4"
matplotlib = "^3.8.3"
transformers = "^4.39.3"
torch = "^2.2.2"
accelerate = "^0.29.1"
ipykernel = "^6.29.4"
pytest = "^8.2.2"
minio = "^7.2.7"
sqlalchemy = "^2.0.31"
psycopg2 = "^2.9.9"
pydantic = "^2.11.2"
pydantic-settings = "^2.8.1"
typing-extensions = "^4.13.1"
python-telegram-bot = "^22.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
